Finished Subtitle Implementation
Fully finalized the advanced subtitle system's basic implementation, including distance based subtitle hiding (which apparently was already built-in with the way I made this, so that was awesome to find out), as well as a default widget behaviour, only requiring the user to implement a widget off of the abstract widget and implement the wanted text widgets wherever they please. Next to do for the subtitle system is extra stuff, like implementing optional name attachment/removal, as well as optional distance/direction handling for impaired users to better play.